--- src/types/address/accPublicKey.spec.ts
@@ -1,59 +1,0 @@
-import { proto, cosmosclient } from '../..';
-import { setBech32Prefix } from './config';
-
-describe('address', () => {
-  it('ed25519', async () => {
-    expect.hasAssertions();
-
-    //set prfix for test
-    setBech32Prefix({
-      accAddr: 'ununifi',
-      accPub: 'ununifipub',
-      valAddr: 'ununifivaloper',
-      valPub: 'ununifivaloperpub',
-      consAddr: 'ununifivalcons',
-      consPub: 'ununifivalconspub',
-    });
-
-    //input mnemonic
-    const mnemonic =
-      'chest flight brain grocery flock elephant gloom gaze diet girl subway again extra spider monitor kiss explain paper beauty ordinary ship dry oxygen shield';
-    const privKey = new proto.cosmos.crypto.ed25519.PrivKey({
-      key: await cosmosclient.generatePrivKeyFromMnemonic(mnemonic),
-    });
-    const pubkey = privKey.pubKey();
-    console.log('accPubEd', pubkey.accPubkey());
-    console.log('toStringEd', typeof pubkey.toString(), pubkey.toString());
-
-    //check
-    //!!! this public key is came from secp256k1 privatekey !!!
-    expect(pubkey.accPubkey()).toBe('ununifipub1addwnpepq0u4zl9r2x4ks82mjetexffsdduruqkmmtmqnx68dfkuy2yr275e53rn0e4');
-  });
-
-  it('secp256k1', async () => {
-    expect.hasAssertions();
-
-    //set prfix for test
-    setBech32Prefix({
-      accAddr: 'ununifi',
-      accPub: 'ununifipub',
-      valAddr: 'ununifivaloper',
-      valPub: 'ununifivaloperpub',
-      consAddr: 'ununifivalcons',
-      consPub: 'ununifivalconspub',
-    });
-
-    //input mnemonic
-    const mnemonic =
-      'chest flight brain grocery flock elephant gloom gaze diet girl subway again extra spider monitor kiss explain paper beauty ordinary ship dry oxygen shield';
-    const privKey = new proto.cosmos.crypto.secp256k1.PrivKey({
-      key: await cosmosclient.generatePrivKeyFromMnemonic(mnemonic),
-    });
-    const pubkey = privKey.pubKey();
-    console.log('accPubSp', pubkey.accPubkey());
-    console.log('toStringSp', typeof pubkey.toString(), pubkey.toString());
-
-    //check
-    expect(pubkey.accPubkey()).toBe('ununifipub1addwnpepq0u4zl9r2x4ks82mjetexffsdduruqkmmtmqnx68dfkuy2yr275e53rn0e4');
-  });
-});
