+++ src/types/crypto/secp256k1.spec.ts
@@ -1,4 +1,5 @@
 import { proto, cosmosclient } from '../..';
+import { setBech32Prefix } from './../address/config';
 import * as crypto from 'crypto';
 import * as secp256k1 from 'secp256k1';
 
@@ -24,5 +25,32 @@ describe('secp256k1', () => {
     const sig = key.sign(msg);
 
     expect(pubkey.verify(msg, sig)).toBeTruthy();
+  });
+
+  it('secp256k1_accPubkey', async () => {
+    expect.hasAssertions();
+
+    //set prfix for test
+    setBech32Prefix({
+      accAddr: 'ununifi',
+      accPub: 'ununifipub',
+      valAddr: 'ununifivaloper',
+      valPub: 'ununifivaloperpub',
+      consAddr: 'ununifivalcons',
+      consPub: 'ununifivalconspub',
+    });
+
+    //input mnemonic
+    const mnemonic =
+      'chest flight brain grocery flock elephant gloom gaze diet girl subway again extra spider monitor kiss explain paper beauty ordinary ship dry oxygen shield';
+    const privKey = new proto.cosmos.crypto.secp256k1.PrivKey({
+      key: await cosmosclient.generatePrivKeyFromMnemonic(mnemonic),
+    });
+    const pubkey = privKey.pubKey();
+    console.log('accPubSp', pubkey.accPubkey());
+    console.log('toStringSp', typeof pubkey.toString(), pubkey.toString());
+
+    //check
+    expect(pubkey.accPubkey()).toBe('ununifipub1addwnpepq0u4zl9r2x4ks82mjetexffsdduruqkmmtmqnx68dfkuy2yr275e53rn0e4');
   });
 });
